+++ packages/react-native-reanimated/src/mappers.native.ts
@@ -2,9 +2,13 @@
 
 import { scheduleOnUI } from 'react-native-worklets';
 
-import type { MapperOutputs, MapperRawInputs } from './commonTypes';
+import type {
+  Mapper,
+  MapperExtractedInputs,
+  MapperOutputs,
+  MapperRawInputs,
+} from './commonTypes';
 import { isSharedValue } from './isSharedValue';
-import type { Mapper, MapperExtractedInputs } from './mappersCommon';
 
 function createMapperRegistry() {
   'worklet';
